Final Commit for V2.2
This will be the final commit to this project for a long time as I am starting to work on Project-BRD V2.3. (Now renamed to BRD-Sharp.)

I primarily just made this commit to have any code local to my machine, on the internet in case of a failure. But hey, at least Project-BRD is getting bigger and better! It is no longer just this indie project, well, I guess it is because it is still being developed by 1 person, but it is starting to have many more features of a fully fleshed out project management tool.

There should be a commit directly after this one where I push what I have of the newer version (so far) to Github.